reduced buttons to 16. updated other classes

+++ src/HID-APIs/GamepadAPI.hpp
@@ -45,12 +45,12 @@ void GamepadAPI::write(void){
 
 
 void GamepadAPI::press(uint8_t b){ 
-	_report.buttons |= (uint32_t)1 << (b - 1); 
+	_report.buttons |= (uint16_t)1 << (b - 1); 
 }
 
 
 void GamepadAPI::release(uint8_t b){ 
-	_report.buttons &= ~((uint32_t)1 << (b - 1)); 
+	_report.buttons &= ~((uint16_t)1 << (b - 1)); 
 }
 
 
@@ -58,7 +58,7 @@ void GamepadAPI::releaseAll(void){
 	memset(&_report, 0x00, sizeof(_report)); 
 }
 
-void GamepadAPI::buttons(uint32_t b){ 
+void GamepadAPI::buttons(uint16_t b){ 
 	_report.buttons = b; 
 }
